+++ .env.template
@@ -4,3 +4,6 @@ FLASK_ENV=development
 
 # Change the following values for local development.
 SECRET_KEY=secret-key
+API_KEY=api-key
+API_TOKEN=api-token
+TRELLO_BOARD_ID=trello-board-id
